move and document ConflictingSystemConnectionMemo -> CaptiveSystemConnectionMemo
