+++ help/en/html/doc/Technical/Names.shtml
@@ -109,6 +109,17 @@
       "P" if you want to. If you have two of them, you can call one "X" and the other "Y". You can
       also use a lower case or upper case letter followed by digits, e.g. "N1" and "n2".</p>
 
+      <p>Note:</p>
+      There are three special cases.  These are JMRI NamedBean types that are not
+      connected to any single system connection, so they always use a particular
+      system letter prefix and type letter regardless of whether that prefix has been configured to a
+      connection:
+      <ul>
+      <li>OBlocks always use OB
+      <li>Transponding tags always use LD
+      <li>RailComm tags always use RD
+      </ul>
+
       <h4>Default Connection Letters</h4>
 
       <p>Note that some of these are placeholders, and have no underlying implementation. (Links
@@ -229,7 +240,7 @@
 
       <h3 id="types">Device type letters</h3>
       Note that some of these are placeholders, and have no underlying implementation. Also, there
-      is no guarantee that any given hardware system will ever implement all of them. 
+      is no guarantee that any given hardware system will ever implement all of them.
       <!-- list of in-use letters from running "grep -A5 -r 'public char typeLetter' java/src | grep return" -->
 
       <ul>
