+++ java/src/jmri/managers/DefaultRailComManager.java
@@ -22,9 +22,8 @@ import org.slf4j.LoggerFactory;
 public class DefaultRailComManager extends DefaultIdTagManager
         implements RailComManager {
 
-    @SuppressWarnings("deprecation") // ConflictingSystemConnectionMemo
     public DefaultRailComManager() {
-        super(new jmri.jmrix.ConflictingSystemConnectionMemo("R", "RailCom")); // NOI18N
+        super(new jmri.jmrix.CaptiveSystemConnectionMemo("R", "RailCom")); // NOI18N
         setInstances();
     }
